+++ packages/clerk-js/src/ui/components/SignIn/SignInStart.tsx
@@ -448,7 +448,10 @@ const InstantPasswordRow = ({ field }: { field?: FormControlState<'password'> })
   useLayoutEffect(() => {
     const intervalId = setInterval(() => {
       if (ref?.current) {
-        const autofilled = window.getComputedStyle(ref.current, ':autofill').animationName === 'onAutoFillStart';
+        const autofilled =
+          window.getComputedStyle(ref.current, ':autofill').animationName === 'onAutoFillStart' ||
+          // https://github.com/facebook/react/issues/1159#issuecomment-1025423604
+          !!ref.current?.matches('*:-webkit-autofill');
         if (autofilled) {
           setAutofilled(autofilled);
           clearInterval(intervalId);